renamed types/default_types -> types/properties
