+++ include/gl/edge_descriptor.hpp
@@ -2,7 +2,6 @@
 
 #include "edge_tags.hpp"
 #include "gl/attributes/force_inline.hpp"
-#include "types/default_types.hpp"
 #include "types/formatter.hpp"
 #include "types/type_traits.hpp"
 #include "types/types.hpp"
