+++ include/gl/types/properties.hpp
