+++ include/gl/vertex_descriptor.hpp
@@ -3,7 +3,7 @@
 #include "edge_tags.hpp"
 #include "gl/attributes/force_inline.hpp"
 #include "gl/impl/impl_tags_decl.hpp"
-#include "types/default_types.hpp"
+#include "types/properties.hpp"
 #include "types/type_traits.hpp"
 #include "types/types.hpp"
